데이터 테이블 2.1 적용중

+++ Assets/Scripts/08.Ui/UiCurrency.cs
@@ -15,6 +15,6 @@ public class UiCurrency : Observer
 
     public override void Notify(Subject subject)
     {
-        textCurrency.text = CurrencyManager.currency[(int)currencyType].ToString();
+        textCurrency.text = CurrencyManager.currency[currencyType].ToString();
     }
 }
